requirements.txt
nicegui>=2.0
SQLAlchemy>=2.0
Pillow
